requirements.txt
# Core dependencies for audio transcription with diarization
# Pinned versions for stability
openai-whisper>=20231117
torch>=2.0.0,<3.0.0
torchaudio>=2.0.0,<3.0.0
librosa>=0.10.0,<0.11.0
soundfile>=0.12.0,<0.13.0

# NeMo toolkit for speaker diarization
nemo_toolkit[asr]>=1.20.0

# Machine learning and clustering
scikit-learn>=1.3.0,<2.0.0
numpy>=1.24.0,<2.0.0

# Progress bars and utilities
tqdm>=4.66.0
python-dotenv>=1.0.0

# Optional but recommended
psutil>=5.9.0  # System resource monitoring

# Web UI
streamlit>=1.28.0  # Web interface for localhost